+++ packages/fighting-design/alert/src/interface.d.ts
@@ -1,6 +1,6 @@
 export type alertType = 'primary' | 'success' | 'danger' | 'warning' | 'default'
 
-export type overflowType = 'hidden' | 'roll' | 'ellipsis' | ''
+export type overflowType = 'hidden' | ''
 
 export interface handleCloseInterface {
   (evt: MouseEvent): void